pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yomibiyori"
version = "0.1.0"
description = "Yomibiyori poetic SNS backend"
authors = [{name = "Yomibiyori"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.29",
  "psycopg[binary]>=3.1.12",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.4.0",
  "pyjwt[crypto]>=2.8.0",
  "python-jose[cryptography]>=3.3.0",
  "requests>=2.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "httpx>=0.26.0",
  "faker>=24.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
